move more code to Package
